+++ pytorch_ext/__init__.py
@@ -3,5 +3,5 @@ from . import util
 from . import objective
 from . import functional
 
-__version__ = "0.7.2"
+__version__ = "0.7.3"
 __author__  = "David Leon (Dawei Leng)"