+++ docs/history.md
@@ -1,4 +1,7 @@
 # History
 
 ## version 0.7.2 [11-5-2019]
-* **NEW**: Add `util.verbose_print` for `print` with verbose level filtering 
+* **NEW**: Add `util.verbose_print()` for `print` with verbose level filtering 
+
+## version 0.7.3 [12-26-2019]
+* **NEW**: Add `util.freeze_module()/unfreeze_module()/get_trainable_parameters()` for training parameters handling